dtext: merge external dtext repo into main repo.
Merge the https://github.com/danbooru/dtext_rb repo into this repo.

Instead of having the DText library live in a separate repo, it now lives in the main Danbooru repo
under `lib/dtext_rb`. This will make development easier by not splitting things into two repos.

The main reason the DText library was in a separate repo was basically because of a technical
limitation of Bundler. The DText library builds a C extension, and the recommended way to do that is
to structure your project as a gem and use rake-compiler to build it [1]. But Bundler really wants
gems to live in a separate Git repo rather than in a subdirectory of the main project. If you put
the gem in a subdirectory, then Bundler won't build the C extension [2]. There's a old issue about
this but it was closed as wontfix [3]. We could build the extension manually and not install it as a
gem, but that would leave the dtext.so library inside `lib/dtext_rb`, which would cause problems
with Docker (we want build artifacts to be separate from the source code so that you can mount code
from the host into the container without it overwriting any prebuilt artifacts in the container).

The way it works now is we basically just build and install the DText gem before running `bundle
install`, rather than relying on `bundle install` to build the DText gem itself. This requires an
extra step in the build process, but it simplifies things overall by not needing a separate repo.

[1]: https://guides.rubygems.org/gems-with-extensions/
[2]: https://bundler.io/v2.5/man/gemfile.5.html#PATH
[3]: https://github.com/rubygems/bundler/issues/1679