dtext: remove pre-generated dtext.cpp file.
Instead of including the pre-generated dtext.cpp file in the repo, generate it when the DText gem is
built. This way we don't have to regenerate and commit it every time we update DText.

This requires you to have Ragel installed to build the gem. If you were building the DText extension
manually before, you'll have to generate dtext.cpp like this:

  ragel -G2 ext/dtext/dtext.cpp.rl -o ext/dtext/dtext.cpp